# views/docs/swagger.yaml
basePath: /v3
definitions:
...
  constants.ATCRating:
    enum:
    - -1
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 7
    - 8
    - 10
    - 11
    - 12
    type: integer
    x-enum-varnames:
    - InactiveRating
    - SuspendedRating
    - ObserverRating
    - Student1Rating
    - Student2Rating
    - Student3Rating
    - ControllerRating
    - SeniorControllerRating
    - InstructorRating
    - SeniorInstructorRating
    - SupervisorRating
    - AdministratorRating
  constants.FacilityID:
    enum:
    - ZAE
    - ZHQ
    - ZZN
    - ZZI
    - ZAB
    - ZAN
    - ZTL
    - ZBW
    - ZAU
    - ZOB
    - ZDV
    - ZFW
    - HCF
    - ZHU
    - ZID
    - ZJX
    - ZKC
    - ZLA
    - ZME
    - ZMA
    - ZMP
    - ZNY
    - ZOA
    - ZLC
    - ZSE
    - ZDC
    type: string
    x-enum-varnames:
    - AcademyFacility
    - HeadquartersFacility
    - NonMemberFacility
    - InactiveFacility
    - AlbuquerqueFacility
    - AnchorageFacility
    - AtlantaFacility
    - BostonFacility
    - ChicagoFacility
    - ClevelandFacility
    - DenverFacility
    - FortWorthFacility
    - HonoluluFacility
    - HoustonFacility
    - IndianapolisFacility
    - JacksonvilleFacility
    - KansasCityFacility
    - LosAngelesFacility
    - MemphisFacility
    - MiamiFacility
    - MinneapolisFacility
    - NewYorkFacility
    - OaklandFacility
    - SaltLakeFacility
    - SeattleFacility
    - WashingtonFacility
...
  faq.Request:
    properties:
      answer:
        example: You should join ZDV.
        type: string
      category:
        enum:
        - membership
        - training
        - technology
        - misc
        example: membership
        type: string
      question:
        example: What ARTCC should I join?
        type: string
    required:
    - answer
    - category
    - question
    type: object
  faq.Response:
    properties:
      answer:
        example: There are no reasons not to join ZDV!
        type: string
      category:
        example: membership
        type: string
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      created_by:
        example: 1293257
        type: integer
      facility:
        allOf:
        - $ref: '#/definitions/constants.FacilityID'
        example: ZDV
      id:
        example: 1
        type: integer
      question:
        example: Why shouldn't I join ZDV?'
        type: string
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      updated_by:
        example: 1293257
        type: integer
    type: object
  feedback.Request:
    properties:
      callsign:
        example: DAL123
        type: string
      comment:
        example: Great work Raaj!
        type: string
      controller_cid:
        example: 1293257
        type: integer
      feedback:
        example: Raaj was the best controller I've ever flown under.
        type: string
      pilot_cid:
        example: 1293257
        type: integer
      position:
        example: DEN_I_APP
        type: string
      rating:
        allOf:
        - $ref: '#/definitions/types.FeedbackRating'
        enum:
        - unsatisfactory
        - poor
        - fair
        - good
        - excellent
        example: good
      status:
        allOf:
        - $ref: '#/definitions/types.StatusType'
        enum:
        - pending
        - approved
        - denied
        example: pending
    required:
    - callsign
    - controller_cid
    - feedback
    - pilot_cid
    - position
    - rating
    - status
    type: object
  feedback.Response:
    properties:
      callsign:
        example: DAL123
        type: string
      comment:
        example: Great work Raaj!
        type: string
      controller_cid:
        example: 1293257
        type: integer
      controller_first_name:
        example: John
        type: string
      controller_last_name:
        example: Doe
        type: string
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      facility:
        allOf:
        - $ref: '#/definitions/constants.FacilityID'
        example: ZDV
      feedback:
        example: Raaj was the best controller I've ever flown under.
        type: string
      id:
        example: 1
        type: integer
      pilot_cid:
        example: 1293257
        type: integer
      position:
        example: DEN_I_APP
        type: string
      rating:
        allOf:
        - $ref: '#/definitions/types.FeedbackRating'
        example: good
      status:
        allOf:
        - $ref: '#/definitions/types.StatusType'
        example: pending
    type: object
...
  rating_change.Request:
    properties:
      new_rating:
        allOf:
        - $ref: '#/definitions/constants.ATCRating'
        example: 2
      old_rating:
        allOf:
        - $ref: '#/definitions/constants.ATCRating'
        example: 1
    required:
    - new_rating
    - old_rating
    type: object
  rating_change.Response:
    properties:
      cid:
        example: 1293257
        type: integer
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      created_by_cid:
        example: 1293257
        type: integer
      id:
        example: 1
        type: integer
      new_rating:
        allOf:
        - $ref: '#/definitions/constants.ATCRating'
        example: 2
      old_rating:
        allOf:
        - $ref: '#/definitions/constants.ATCRating'
        example: 1
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
...
  types.FeedbackRating:
    enum:
    - unsatisfactory
    - poor
    - fair
    - good
    - excellent
    type: string
    x-enum-varnames:
    - Unsatisfactory
    - Poor
    - Fair
    - Good
    - Excellent
...
  types.StatusType:
    enum:
    - ""
    - pending
    - accepted
    - rejected
    type: string
    x-enum-varnames:
    - All
    - Pending
    - Accepted
    - Rejected
...
  utils.ErrResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
info:
  contact:
    email: vatusa6@vatusa.net
    name: VATUSA Support
    url: http://www.swagger.io/support
  description: VATUSAs public API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VATUSA API
  version: "0.1"
paths:
...
  /facility/{FacilityID}/faq:
    get:
      consumes:
      - application/json
      description: List all FAQs
      parameters:
      - description: Facility ID
        in: query
        name: FacilityID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/faq.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: List all FAQs
      tags:
      - faq
    post:
      consumes:
      - application/json
      description: Create a new FAQ
      parameters:
      - description: FAQ
        in: body
        name: faq
        required: true
        schema:
          $ref: '#/definitions/faq.Request'
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/faq.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create a new FAQ
      tags:
      - faq
  /facility/{FacilityID}/faq/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a FAQ
      parameters:
      - description: FAQ ID
        in: path
        name: id
        required: true
        type: string
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Delete a FAQ
      tags:
      - faq
    patch:
      consumes:
      - application/json
      description: Patch a FAQ
      parameters:
      - description: FAQ ID
        in: path
        name: id
        required: true
        type: string
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      - description: FAQ
        in: body
        name: faq
        required: true
        schema:
          $ref: '#/definitions/faq.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/faq.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Patch a FAQ
      tags:
      - faq
    put:
      consumes:
      - application/json
      description: Update a FAQ
      parameters:
      - description: FAQ ID
        in: path
        name: id
        required: true
        type: string
      - description: Facility ID
        in: path
        name: FacilityID
        required: true
        type: string
      - description: FAQ
        in: body
        name: faq
        required: true
        schema:
          $ref: '#/definitions/faq.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/faq.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update a FAQ
      tags:
      - faq
  /facility/{FacilityID}/feedback:
    get:
      consumes:
      - application/json
      description: List feedback entries
      parameters:
      - description: Facility ID
        in: path
        name: facility
        required: true
        type: string
      - description: CID
        in: query
        name: cid
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feedback.Response'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: List feedback entries
      tags:
      - feedback
    post:
      consumes:
      - application/json
      description: Create a new feedback entry
      parameters:
      - description: Facility ID
        in: path
        name: facility
        required: true
        type: string
      - description: Feedback Entry
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/feedback.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/feedback.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create a new feedback entry
      tags:
      - feedback
  /facility/{FacilityID}/feedback/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a feedback entry
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: integer
      - description: Facility ID
        in: path
        name: facility
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Delete a feedback entry
      tags:
      - feedback
    patch:
      consumes:
      - application/json
      description: Patch a feedback entry
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: integer
      - description: Facility ID
        in: path
        name: facility
        required: true
        type: string
      - description: Feedback Entry
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/feedback.Request'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Patch a feedback entry
      tags:
      - feedback
    put:
      consumes:
      - application/json
      description: Update a feedback entry
      parameters:
      - description: Feedback ID
        in: path
        name: id
        required: true
        type: integer
      - description: Facility ID
        in: path
        name: facility
        required: true
        type: string
      - description: Feedback Entry
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/feedback.Request'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update a feedback entry
      tags:
      - feedback
...
  /user/{cid}/feedback:
    get:
      consumes:
      - application/json
      description: Get feedback entries for a user
      parameters:
      - description: CID
        in: path
        name: CID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/feedback.Response'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get accepted feedback entries for a user
      tags:
      - feedback
...
  /user/{cid}/rating-change:
    get:
      consumes:
      - application/json
      description: List rating changes
      parameters:
      - description: CID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rating_change.Response'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: List rating changes
      tags:
      - rating-change
    post:
      consumes:
      - application/json
      description: Create a new rating change
      parameters:
      - description: CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Rating Change
        in: body
        name: rating_change
        required: true
        schema:
          $ref: '#/definitions/rating_change.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rating_change.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create a new rating change
      tags:
      - rating-change
  /user/{cid}/rating-change/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a rating change
      parameters:
      - description: Rating Change ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Delete a rating change
      tags:
      - rating-change
    patch:
      consumes:
      - application/json
      description: Patch a rating change
      parameters:
      - description: CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Rating Change ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rating Change
        in: body
        name: rating_change
        required: true
        schema:
          $ref: '#/definitions/rating_change.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rating_change.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Patch a rating change
      tags:
      - rating-change
    put:
      consumes:
      - application/json
      description: Update a rating change
      parameters:
      - description: CID
        in: path
        name: cid
        required: true
        type: integer
      - description: Rating Change ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rating Change
        in: body
        name: rating_change
        required: true
        schema:
          $ref: '#/definitions/rating_change.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rating_change.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update a rating change
      tags:
      - rating-change
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"